Check out submodules after updating
This commit adds the --checkout flag to all invocations of 'git
submodule update --init --recursive'. This ensures that all submodules
are checked out, even if they have the "update = none" directive in the
.gitmodules file. This fixes a bug where checking out a commit would
fail with an error message indicating that a subdirectory is 'not a git
repository'.

This commit also removes `--recurse-submodules` as an option to 'git
checkout'. This is because we want to be able to check out a particular
commit (at the head of a PR) and then separately run 'git submodule
update' with the right flags. '--recurse-submodule' is a convenience
option that updates the submodules, but doesn't offer as many options as
doing a separate checkout followed by a submodule update.